% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aversion_data}
\alias{aversion_data}
\title{Example assay data}
\format{
A data frame with 231 rows and 4 variables:
\describe{
  \item{date}{date the assay was ran in ymd format}
  \item{worm_strain}{the worm strain being tested}
  \item{compound}{the compound that was used to make the ring}
  \item{attempts}{the number of attempts (out of 6) that the worm required to escape the ring}
  ...
}
}
\source{
{the Zamanian Lab data cloud}
}
\usage{
aversion_data
}
\description{
Data for a few ring-based aversion experiments that Kendra ran
with transgenic worms
}
\keyword{datasets}
