% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_zlab.R
\name{theme_nw}
\alias{theme_nw}
\title{Nic's theme}
\usage{
theme_nw(base_size = 16, base_family = "helvetica")
}
\arguments{
\item{base_size}{An integer.}

\item{base_family}{A font}
}
\value{

}
\description{
Nic's theme
}
\examples{
library("ggplot2")

box <- ggplot(data = iris, aes(x = Species, y = Sepal.Length)) +
  geom_boxplot(aes(fill = Species)) +
  ylab("Sepal Length") + ggtitle("Iris Boxplot") +
  stat_summary(fun = mean, geom = "point", shape = 5, size = 4) +
  theme_nw2()
box
}
