% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_zlab.R
\name{theme_zlab}
\alias{theme_zlab}
\title{Implement the base Zamanian Lab theme for ggplot2 plots}
\usage{
theme_zlab(base_size = 16, base_family = "helvetica")
}
\arguments{
\item{base_size}{An integer.}

\item{base_family}{An integer.}
}
\value{

}
\description{
Implement the base Zamanian Lab theme for ggplot2 plots
}
\examples{
library("ggplot2")

box <- ggplot(data = iris, aes(x = Species, y = Sepal.Length)) +
  geom_boxplot(aes(fill = Species)) +
  ylab("Sepal Length") + ggtitle("Iris Boxplot") +
  stat_summary(fun = mean, geom = "point", shape = 5, size = 4) +
  theme_zlab() +
  scale_fill_transgenes()
box
}
