% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_zlab.R
\name{transgenes_pal}
\alias{transgenes_pal}
\title{Color palette for C. elegans tissue-specific transgenesis}
\usage{
transgenes_pal()
}
\value{

}
\description{
Color palette for C. elegans tissue-specific transgenesis
}
\examples{
library("ggplot2")

box <- ggplot(data = iris, aes(x = Species, y = Sepal.Length)) +
  geom_boxplot(aes(fill = Species)) +
  ylab("Sepal Length") + ggtitle("Iris Boxplot") +
  stat_summary(fun = mean, geom = "point", shape = 5, size = 4) +
  theme_zlab() +
  scale_fill_transgenes()
box
}
